Update in response to Go release.2011-01-20 (removal of the 'float' type), the SDL-sound code now uses 'float32' instead of 'float'

+++ src/gospeccy.go
@@ -73,7 +73,7 @@ type SDLRenderer struct {
 	consoleW, consoleH, consoleH_2, width, height uint16
 	appSurface                                    *sdl.Surface
 	speccySurface, cliSurface                     SDLSurfaceAccessor
-	fps                                           float
+	fps                                           float32
 	toggling                                      bool
 	animationDuration                             int64
 }
@@ -385,7 +385,7 @@ func main() {
 	help := flag.Bool("help", false, "Show usage")
 	scale2x := flag.Bool("2x", false, "2x display scaler")
 	fullscreen := flag.Bool("fullscreen", false, "Fullscreen (enable 2x scaler by default)")
-	fps := flag.Float("fps", spectrum.DefaultFPS, "Frames per second")
+	fps := flag.Float64("fps", spectrum.DefaultFPS, "Frames per second")
 	sound := flag.Bool("sound", true, "Enable or disable sound")
 	acceleratedLoad := flag.Bool("accelerated-load", false, "Enable or disable accelerated tapes loading")
 	verbose := flag.Bool("verbose", false, "Enable debugging messages")
@@ -474,7 +474,7 @@ func main() {
 	go r.loop()
 
 	// Set the FPS
-	speccy.CommandChannel <- spectrum.Cmd_SetFPS{*fps}
+	speccy.CommandChannel <- spectrum.Cmd_SetFPS{float32(*fps)}
 
 	// Process command line argument. Load the given program (if any)
 	if flag.Arg(0) != "" {
